Improve typography, sample output, & documentation
Typography:
- The spacing in sets generated by the `\set` command has been reduced.
- The `tracking` option has been added to `microtype`.
- `\frenchspacing` is now invoked through `babel` instead of directly.

Sample output:
- The template `enumerate` environment is now generated using
`blindtext`. Math is now included within the sample text so that it is
more representative of the types of documents likely to be produced
using this template. The sample text now extends beyond a single page to
let potential users see the margins implemented by this template.
- Example values have been added to the header.

Documentation:
- Explanations have been added to some undocumented commands.